Update jackson dependencies to 2.7.9.5 and set scope to provided
Version 2.7.9.5 resolves a number of security vulnerabilities in
jackson-databind. Changing the dependency scope to 'provided' forces
the Status Servlet extension to rely on the server's (e.g., Directory
Server, DG) Jackson libraries at runtime, so in practice, this extension
should not be vulnerable to future Jackson security issues as long as
the server is not.